break up the formal integral components
