+++ tardis/spectrum/base.py
@@ -4,7 +4,7 @@ import numpy as np
 from astropy import units as u
 
 from tardis.io.hdf_writer_mixin import HDFWriterMixin
-from tardis.spectrum.formal_integral.formal_integral import IntegrationError
+from tardis.spectrum.formal_integral.base import IntegrationError
 from tardis.spectrum.spectrum import TARDISSpectrum
 from tardis.util.base import (
     quantity_linspace,
