+++ tardis/spectrum/formal_integral/tests/test_numba_formal_integral.py
@@ -4,7 +4,9 @@ import pytest
 
 from tardis import constants as c
 from tardis.model.geometry.radial1d import NumbaRadial1DGeometry
-import tardis.spectrum.formal_integral.formal_integral as formal_integral
+from tardis.spectrum.formal_integral.base import C_INV
+from tardis.spectrum.formal_integral.formal_integral import FormalIntegrator
+import tardis.spectrum.formal_integral.formal_integral_numba as formal_integral_numba
 from tardis.util.base import intensity_black_body
 
 
@@ -17,7 +19,7 @@ from tardis.util.base import intensity_black_body
     ],
 )
 def test_intensity_black_body(nu, temperature):
-    func = formal_integral.intensity_black_body
+    func = formal_integral_numba.intensity_black_body
     actual = func(nu, temperature)
     print(actual, type(actual))
     expected = intensity_black_body(nu, temperature)
@@ -26,7 +28,7 @@ def test_intensity_black_body(nu, temperature):
 
 @pytest.mark.parametrize("N", (1e2, 1e3, 1e4, 1e5))
 def test_trapezoid_integration(N):
-    func = formal_integral.trapezoid_integration
+    func = formal_integral_numba.trapezoid_integration
     h = 1.0
     N = int(N)
     data = np.random.random(N)
@@ -72,7 +74,7 @@ def calculate_z(r, p):
 @pytest.mark.parametrize("p", [0.0, 0.5, 1.0])
 def test_calculate_z(formal_integral_geometry, time_explosion, p):
 
-    func = formal_integral.calculate_z
+    func = formal_integral_numba.calculate_z
     inv_t = 1.0 / time_explosion
     size = len(formal_integral_geometry.r_outer)
     r_outer = formal_integral_geometry.r_outer
@@ -82,7 +84,7 @@ def test_calculate_z(formal_integral_geometry, time_explosion, p):
         if p >= r:
             assert actual == 0
         else:
-            desired = np.sqrt(r * r - p * p) * formal_integral.C_INV * inv_t
+            desired = np.sqrt(r * r - p * p) * C_INV * inv_t
             ntest.assert_almost_equal(actual, desired)
 
 
@@ -92,8 +94,8 @@ def test_populate_z_photosphere(formal_integral_geometry, time_explosion, p):
     Test the case where p < r[0]
     That means we 'hit' all shells from inside to outside.
     """
-    integrator = formal_integral.FormalIntegrator(time_explosion, None, None)
-    func = formal_integral.populate_z
+    integrator = FormalIntegrator(time_explosion, None, None)
+    func = formal_integral_numba.populate_z
     size = len(formal_integral_geometry.r_outer)
     r_inner = formal_integral_geometry.r_inner
     r_outer = formal_integral_geometry.r_outer
@@ -115,8 +117,8 @@ def test_populate_z_shells(formal_integral_geometry, time_explosion, p):
     """
     Test the case where p > r[0]
     """
-    integrator = formal_integral.FormalIntegrator(time_explosion, None, None)
-    func = formal_integral.populate_z
+    integrator = FormalIntegrator(time_explosion, None, None)
+    func = formal_integral_numba.populate_z
 
     size = len(formal_integral_geometry.r_inner)
     r_inner = formal_integral_geometry.r_inner
@@ -165,7 +167,7 @@ def test_populate_z_shells(formal_integral_geometry, time_explosion, p):
 )
 def test_calculate_p_values(N):
     r = 1.0
-    func = formal_integral.calculate_p_values
+    func = formal_integral_numba.calculate_p_values
 
     expected = r / (N - 1) * np.arange(0, N, dtype=np.float64)
     actual = np.zeros_like(expected, dtype=np.float64)
